Fix hostname check in dot_zshrc.tmpl
The Chezmoi hostname check uses entries from /etc/hosts, which is
altered on my work computer. The "is this my work computer" check is now
expressed as "is this a Mac running macOS" and "is my username
igor.stefanco" which ought to be specific enough. Regular hostname
should be fine on for the mac mini.